// frontend/src/interfaces/ProfileInterfaces.ts
export interface MotherProfile {
  id: string;
  firstname: string;
  lastname: string;
  username: string;
  email: string;
  profile_pic: string;
  date_of_birth: string; 
  phone_number: string;
  address: string;
  city: string;
  country: string;
  number_of_children: number;
  blood_type: string;
  account_created_at: string; 
}

export interface MotherProfileUpdate {
  
}